backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djoser==2.2.3
django-cors-headers==4.4.0
drf-spectacular==0.27.2
django-environ==0.11.2
channels==4.1.0
daphne==4.1.2
redis==5.1.1